+++ scripts/mods/BetterInventory/BetterInventory_layout_blueprints.lua
@@ -92,7 +92,6 @@ local configure_new_item_highlight = Cards.configure_new_item_highlight
 local add_custom_content_passes = Cards.add_custom_content_passes
 local grid_weapon_name_font_size = Cards.grid_weapon_name_font_size
 local configure_card_content = Cards.configure_card_content
-local configure_compound_weapon_icon_fallback = Cards.configure_compound_weapon_icon_fallback
 
 local function content_is_curio(card_content)
 	local cached = card_content and card_content.better_inventory_is_curio
@@ -918,7 +917,6 @@ Blueprints.configure_item_blueprint = function(mod, item_blueprint, grid_width, 
 	set_height(pass_by_style_id(pass_template, "inner_highlight"), card_height)
 
 	configure_card_content(mod, item_blueprint, configuration)
-	configure_compound_weapon_icon_fallback(item_blueprint, resolved_columns)
 	ImageLayout.apply_blueprint(mod, item_blueprint, configuration, resolved_columns)
 
 	return item_size